ajout suppression + update publication all

+++ back_end/routes/posts.js
@@ -18,7 +18,7 @@ router.post("/actualites", async (req, res) => {
 router.put("/actualites/:id", async (req, res) => {
     try {
         const post = await Post.findById(req.params.id)
-        // if(post.pseudo === req.body.pseudo) {
+        if(post.pseudo === req.body.pseudo) {
 
             try {
                 const updatedPost = await Post.findByIdAndUpdate(req.params.id, {
@@ -29,9 +29,9 @@ router.put("/actualites/:id", async (req, res) => {
             } catch (err) {
                 res.status(500).json(err)
             }
-        // } else {
-        //     res.status(401).json("Vous ne pouvez pas mettre à jour cette publication ! ")
-        // }
+        } else {
+            res.status(401).json("Vous ne pouvez pas mettre à jour cette publication ! ")
+        }
     } catch (err) {
         res.status(500).json(err)
     }
@@ -41,7 +41,7 @@ router.put("/actualites/:id", async (req, res) => {
 router.delete("/actualites/:id", async (req, res) => {
     try {
         const post = await Post.findById(req.params.id)
-        // if(post.pseudo === req.body.pseudo) {
+        if(post.pseudo === req.body.pseudo) {
 
             try {
                await post.delete()
@@ -49,9 +49,9 @@ router.delete("/actualites/:id", async (req, res) => {
             } catch (err) {
                 res.status(500).json(err)
             }
-        // } else {
-        //     res.status(401).json("Vous ne pouvez pas supprimer cette publication ! ")
-        // }
+        } else {
+            res.status(401).json("Vous ne pouvez pas supprimer cette publication ! ")
+        }
     } catch (err) {
         res.status(500).json(err)
     }
